Displayクラスをさらに別クラスに切り分けた
